refactor: refactoring PostListTemplate
