+++ src/templates/PostsListTemplate/constants/index.ts
@@ -1,0 +1,1 @@
+export const EXTERNAL_API_URL = "https://gatsby-related.free.beeceptor.com/";
