+++ src/templates/PostsListTemplate/types/index.ts
@@ -1,0 +1,5 @@
+export type RelatedArticle = {
+  id: string;
+  title: string;
+  url: string;
+};
