+++ src/types/index.ts
@@ -37,9 +37,3 @@ export type ContentfulPostNode = {
 export type ContentfulPostData = {
   contentfulGatsbyBlog: ContentfulPostNode | null;
 };
-
-export type RelatedArticle = {
-  id: string;
-  title: string;
-  url: string;
-};
